--- packages/react-aria-widgets/src/Accordion/utils.ts
@@ -1,9 +1,0 @@
-/**
- * Gets the ID of an accordion panel based on the accordion's ID.
- *
- * @param {string} id
- * @returns {string}
- */
-export function getPanelId(id: string): string {
-  return `${id}-panel`;
-}
