+++ packages/react-aria-widgets/src/types.ts
@@ -1,0 +1,9 @@
+import { Validator } from 'prop-types';
+
+export interface Props {
+  [key: string]: any;
+}
+
+export interface ValidatorWithRequired<T> extends Validator<T> {
+  isRequired?: Validator<T>;
+}
